+++ app/Http/Controllers/KategoriController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\InsertKategoriRequest;
 use App\Models\Data\Kategori;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -12,36 +13,72 @@ class KategoriController extends Controller
     //Menyimpan Kategori ke DB
     public function insert(InsertKategoriRequest $request)
     {
-        $merged = $request->merge(['modified_by' => Auth::id()])->toArray();
-        Kategori::create($merged);
+        $user = User::find(auth()->id());
+        $permissions = $user->getPermissionsViaRoles();
+        $permissionExist = collect($permissions)->first(function ($item) {
+            return $item->name === 'Membuat Kategori';
+        });
+        if ($permissionExist) {
+            $merged = $request->merge(['modified_by' => Auth::id()])->toArray();
+            Kategori::create($merged);
 
-        return redirect()->intended(route('menu-kategori'));
+            return redirect()->intended(route('menu-kategori'));
+        } else {
+            abort(403, 'USER DOES NOT HAVE THE RIGHT PERMISSION');
+        }
     }
 
     //Menampilkan Detail Kategori
     public function show($id)
     {
-        $kategori = Kategori::find($id);
-        return [
-            'status' => 200,
-            $kategori,
-        ];
+        $user = User::find(auth()->id());
+        $permissions = $user->getPermissionsViaRoles();
+        $permissionExist = collect($permissions)->first(function ($item) {
+            return $item->name === 'Melihat Detail Kategori';
+        });
+        if ($permissionExist) {
+            $kategori = Kategori::find($id);
+            return [
+                'status' => 200,
+                $kategori,
+            ];
+        } else {
+            abort(403, 'USER DOES NOT HAVE THE RIGHT PERMISSION');
+        }
     }
 
     //Update Kategori
     public function update(InsertKategoriRequest $request, $id)
     {
-        $merged = $request->merge(['modified_by' => Auth::id()])->toArray();
-        Kategori::find($id)->update($merged);
+        $user = User::find(auth()->id());
+        $permissions = $user->getPermissionsViaRoles();
+        $permissionExist = collect($permissions)->first(function ($item) {
+            return $item->name === 'Mengubah Data Kategori';
+        });
+        if ($permissionExist) {
+            $merged = $request->merge(['modified_by' => Auth::id()])->toArray();
+            Kategori::find($id)->update($merged);
 
-        return redirect()->intended(route('menu-kategori'));
+            return redirect()->intended(route('menu-kategori'));
+        } else {
+            abort(403, 'USER DOES NOT HAVE THE RIGHT PERMISSION');
+        }
     }
 
     //Menghapus Kategori
     public function delete($id)
     {
-        Kategori::destroy($id);
+        $user = User::find(auth()->id());
+        $permissions = $user->getPermissionsViaRoles();
+        $permissionExist = collect($permissions)->first(function ($item) {
+            return $item->name === 'Menghapus Kategori';
+        });
+        if ($permissionExist) {
+            Kategori::destroy($id);
 
-        return redirect()->intended(route('menu-kategori'));
+            return redirect()->intended(route('menu-kategori'));
+        } else {
+            abort(403, 'USER DOES NOT HAVE THE RIGHT PERMISSION');
+        }
     }
 }
