+++ app/Http/Controllers/PaketDepositController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\InsertPaketDepositRequest;
 use App\Models\Paket\PaketDeposit;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -11,33 +12,69 @@ class PaketDepositController extends Controller
 {
     public function insert(InsertPaketDepositRequest $request)
     {
-        $merged = $request->merge(['modified_by' => Auth::id()])->toArray();
-        PaketDeposit::create($merged);
+        $user = User::find(auth()->id());
+        $permissions = $user->getPermissionsViaRoles();
+        $permissionExist = collect($permissions)->first(function ($item) {
+            return $item->name === 'Membuat Paket Deposit';
+        });
+        if ($permissionExist) {
+            $merged = $request->merge(['modified_by' => Auth::id()])->toArray();
+            PaketDeposit::create($merged);
 
-        return redirect()->intended(route('menu-paket-deposit'));
+            return redirect()->intended(route('menu-paket-deposit'));
+        } else {
+            abort(403, 'USER DOES NOT HAVE THE RIGHT PERMISSION');
+        }
     }
 
     public function show($id)
     {
-        $paket_deposit = PaketDeposit::find($id);
-        return [
-            'status' => 200,
-            $paket_deposit,
-        ];
+        $user = User::find(auth()->id());
+        $permissions = $user->getPermissionsViaRoles();
+        $permissionExist = collect($permissions)->first(function ($item) {
+            return $item->name === 'Melihat Detail Paket Deposit';
+        });
+        if ($permissionExist) {
+            $paket_deposit = PaketDeposit::find($id);
+            return [
+                'status' => 200,
+                $paket_deposit,
+            ];
+        } else {
+            abort(403, 'USER DOES NOT HAVE THE RIGHT PERMISSION');
+        }
     }
 
     public function update(InsertPaketDepositRequest $request, $id)
     {
-        $merged = $request->merge(['modified_by' => Auth::id()])->toArray();
-        PaketDeposit::find($id)->update($merged);
+        $user = User::find(auth()->id());
+        $permissions = $user->getPermissionsViaRoles();
+        $permissionExist = collect($permissions)->first(function ($item) {
+            return $item->name === 'Mengubah Data Paket Deposit';
+        });
+        if ($permissionExist) {
+            $merged = $request->merge(['modified_by' => Auth::id()])->toArray();
+            PaketDeposit::find($id)->update($merged);
 
-        return redirect()->intended(route('menu-paket-deposit'));
+            return redirect()->intended(route('menu-paket-deposit'));
+        } else {
+            abort(403, 'USER DOES NOT HAVE THE RIGHT PERMISSION');
+        }
     }
 
     public function delete($id)
     {
-        PaketDeposit::destroy($id);
+        $user = User::find(auth()->id());
+        $permissions = $user->getPermissionsViaRoles();
+        $permissionExist = collect($permissions)->first(function ($item) {
+            return $item->name === 'Menghapus Paket Deposit';
+        });
+        if ($permissionExist) {
+            PaketDeposit::destroy($id);
 
-        return redirect()->intended(route('menu-paket-deposit'));
+            return redirect()->intended(route('menu-paket-deposit'));
+        } else {
+            abort(403, 'USER DOES NOT HAVE THE RIGHT PERMISSION');
+        }
     }
 }
