+++ app/Http/Controllers/ParfumController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\InsertParfumRequest;
 use App\Models\Data\Parfum;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -11,33 +12,69 @@ class ParfumController extends Controller
 {
     public function insert(InsertParfumRequest $request)
     {
-        $merged = $request->merge(['modified_by' => Auth::id()])->toArray();
-        Parfum::create($merged);
+        $user = User::find(auth()->id());
+        $permissions = $user->getPermissionsViaRoles();
+        $permissionExist = collect($permissions)->first(function ($item) {
+            return $item->name === 'Membuat Parfum';
+        });
+        if ($permissionExist) {
+            $merged = $request->merge(['modified_by' => Auth::id()])->toArray();
+            Parfum::create($merged);
 
-        return redirect()->intended(route('menu-parfum'));
+            return redirect()->intended(route('menu-parfum'));
+        } else {
+            abort(403, 'USER DOES NOT HAVE THE RIGHT PERMISSION');
+        }
     }
 
     public function show($id)
     {
-        $parfum = Parfum::find($id);
-        return [
-            'status' => 200,
-            $parfum,
-        ];
+        $user = User::find(auth()->id());
+        $permissions = $user->getPermissionsViaRoles();
+        $permissionExist = collect($permissions)->first(function ($item) {
+            return $item->name === 'Melihat Detail Parfum';
+        });
+        if ($permissionExist) {
+            $parfum = Parfum::find($id);
+            return [
+                'status' => 200,
+                $parfum,
+            ];
+        } else {
+            abort(403, 'USER DOES NOT HAVE THE RIGHT PERMISSION');
+        }
     }
 
     public function update(InsertParfumRequest $request, $id)
     {
-        $merged = $request->merge(['modified_by' => Auth::id()])->toArray();
-        Parfum::find($id)->update($merged);
+        $user = User::find(auth()->id());
+        $permissions = $user->getPermissionsViaRoles();
+        $permissionExist = collect($permissions)->first(function ($item) {
+            return $item->name === 'Mengubah Data Parfum';
+        });
+        if ($permissionExist) {
+            $merged = $request->merge(['modified_by' => Auth::id()])->toArray();
+            Parfum::find($id)->update($merged);
 
-        return redirect()->intended(route('menu-parfum'));
+            return redirect()->intended(route('menu-parfum'));
+        } else {
+            abort(403, 'USER DOES NOT HAVE THE RIGHT PERMISSION');
+        }
     }
 
     public function delete($id)
     {
-        Parfum::destroy($id);
+        $user = User::find(auth()->id());
+        $permissions = $user->getPermissionsViaRoles();
+        $permissionExist = collect($permissions)->first(function ($item) {
+            return $item->name === 'Menghapus Parfum';
+        });
+        if ($permissionExist) {
+            Parfum::destroy($id);
 
-        return redirect()->intended(route('menu-parfum'));
+            return redirect()->intended(route('menu-parfum'));
+        } else {
+            abort(403, 'USER DOES NOT HAVE THE RIGHT PERMISSION');
+        }
     }
 }
